2023 jun 06 1245

+++ src/utils.rs
@@ -7,7 +7,13 @@ use libc::EINVAL;
 use windows::Win32::Networking::WinSock;
 use windows::Win32::Networking::WinSock::{socklen_t, ADDRESS_FAMILY, SOCKADDR};
 
-pub fn copy_bytes(dest: &mut [u8], dest_offset: i32, src: &[u8], src_offset: usize, count: i32) {
+pub fn copy_bytes(
+    dest: &mut [u8],
+    dest_offset: usize,
+    src: &[u8],
+    src_offset: usize,
+    count: usize,
+) {
     for i in 0..count {
         dest[dest_offset + i] = src[src_offset + i]
     }
